+++ solver/src/main/java/org/chocosolver/util/objects/setDataStructures/swapList/Set_Std_Swap2.java
@@ -9,12 +9,8 @@
  */
 package org.chocosolver.util.objects.setDataStructures.swapList;
 
-import gnu.trove.list.array.TIntArrayList;
 import org.chocosolver.memory.IEnvironment;
 import org.chocosolver.memory.IStateInt;
-import org.chocosolver.util.objects.setDataStructures.AbstractSet;
-import org.chocosolver.util.objects.setDataStructures.ISetIterator;
-import org.chocosolver.util.objects.setDataStructures.SetType;
 
 /**
  * Set of integers based on BipartiteSet implementation for small sets (arraylist inside)
@@ -22,16 +18,13 @@ import org.chocosolver.util.objects.setDataStructures.SetType;
  *
  * @author : Charles Prud'homme, Jean-Guillaume FAGES (fix remove)
  */
-public class Set_Std_Swap2 extends AbstractSet {
+public class Set_Std_Swap2 extends Set_Swap2 {
 
 	//***********************************************************************************
 	// VARIABLES
 	//***********************************************************************************
 
-    protected IStateInt size;
-    protected TIntArrayList values;
-    private final ISetIterator iter = newIterator();
-
+    private final IStateInt size;
 
     //***********************************************************************************
 	// CONSTRUCTOR
@@ -42,8 +35,8 @@ public class Set_Std_Swap2 extends AbstractSet {
 	 * @param e backtracking environment
 	 */
 	public Set_Std_Swap2(IEnvironment e){
+        super();
 		size = e.makeInt(0);
-        values = new TIntArrayList(4);
 	}
 
 	//***********************************************************************************
@@ -51,107 +44,17 @@ public class Set_Std_Swap2 extends AbstractSet {
 	//***********************************************************************************
 
     @Override
-    public boolean add(int element) {
-        if(!contains(element)){
-            int pos = size.add(1);
-            values.insert(pos - 1, element);
-            notifyObservingElementAdded(element);
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public boolean remove(int element) {
-        int pos = values.indexOf(element);
-        int s = size();
-        if(pos > -1 && pos < s){
-            iter.notifyRemoving(element);
-            s--;
-            int t = values.get(s);
-            values.set(pos, t);
-            values.set(s, element);
-            size.add(-1);
-            notifyObservingElementRemoved(element);
-            return true;
-        }else return false;
-    }
-
-    @Override
-    public boolean contains(int element) {
-        int pos = values.indexOf(element);
-        return pos > -1 && pos < size();
-    }
-
-    @Override
     public int size() {
         return size.get();
     }
 
     @Override
-    public void clear() {
-        size.set(0);
-        notifyObservingCleared();
+    protected void setSize(int s) {
+        size.set(s);
     }
 
     @Override
-    public int min() {
-        if(isEmpty()) throw new IllegalStateException("cannot find minimum of an empty set");
-        int min = values.get(0);
-        for(int i = 1; i< size(); i++){
-            if(min > values.get(i)){
-                min = values.get(i);
-            }
-        }
-        return min;
-    }
-
-    @Override
-    public int max() {
-        if(isEmpty()) throw new IllegalStateException("cannot find maximum of an empty set");
-        int max = values.get(0);
-        for(int i = 1; i< size(); i++){
-            if(max < values.get(i)){
-                max = values.get(i);
-            }
-        }
-        return max;
-    }
-
-    @Override
-    public SetType getSetType() {
-        return SetType.SMALLBIPARTITESET;
-    }
-
-    @Override
-    public ISetIterator iterator(){
-        iter.reset();
-        return iter;
-    }
-
-    @Override
-    public ISetIterator newIterator(){
-        return new ISetIterator() {
-            private int idx;
-            @Override
-            public void reset() {
-                idx = 0;
-            }
-            @Override
-            public void notifyRemoving(int item) {
-                if(idx>0 && item == values.get(idx-1)){
-                    idx--;
-                }
-            }
-            @Override
-            public boolean hasNext() {
-                return idx < size();
-            }
-            @Override
-            public int nextInt() {
-                idx ++;
-                return values.get(idx-1);
-            }
-        };
+    protected void addSize(int delta) {
+        size.add(delta);
     }
 }